requirements.txt
google-generativeai
python-dotenv
pandas
pytest
requests
fastapi
uvicorn
pyttsx3
requests
python-dotenv
pandas
pytest